issues/2 :recycle: Divide interface and implementation file in UseCase
