+++ Assets/Scripts/Domain/UseCase/IFlowUseCase.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 6e57a1a47be2c4a928b07b43ef8bce66
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
