+++ Assets/Scripts/Domain/UseCase/ModelFlowUseCase.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 85c7932168e834e5a9791d04f2786199
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
